speed up tests by consolidation
