+++ test/test_operations_solve.jl
@@ -1,3 +1,6 @@
+import CSV
+import Plots
+
 devices = Dict{Symbol, DeviceModel}(:Generators => DeviceModel(ThermalStandard, ThermalDispatch),
                                     :Loads =>  DeviceModel(PowerLoad, StaticPowerLoad))
 branches = Dict{Symbol, DeviceModel}(:L => DeviceModel(Line, StaticLine),
@@ -8,7 +11,7 @@ services = Dict{Symbol, ServiceModel}()
 @testset "Solving ED with CopperPlate" begin
     template = OperationsProblemTemplate(CopperPlatePowerModel, devices, branches, services);
     parameters_value = [true, false]
-    systems = [c_sys5, c_sys14]
+    systems = [c_sys5] #c_sys14
     test_results = Dict{System, Float64}(c_sys5 => 240000.0,
                                              c_sys14 => 142000.0)
     for sys in systems, p in parameters_value
@@ -24,7 +27,7 @@ end
 @testset "Solving ED with PTDF Models" begin
     template = OperationsProblemTemplate(StandardPTDFModel, devices, branches, services);
     parameters_value = [true, false]
-    systems = [c_sys5, c_sys14, c_sys14_dc]
+    systems = [c_sys5, c_sys14_dc] #c_sys14
     PTDF_ref = Dict{System, PTDF}(c_sys5 => PTDF5, c_sys14 => PTDF14, c_sys14_dc => PTDF14_dc)
     test_results = Dict{System, Float64}(c_sys5 => 340000.0,
                                              c_sys14 => 142000.0,
@@ -40,7 +43,7 @@ end
 end
 
 @testset "Solving ED With PowerModels with loss-less convex models" begin
-    systems = [c_sys5, c_sys14, c_sys14_dc]
+    systems = [c_sys5, c_sys14_dc] #c_sys14
     parameters_value = [true, false]
     networks = [DCPPowerModel,
                 NFAPowerModel]
@@ -61,7 +64,7 @@ end
 end
 
 @testset "Solving ED With PowerModels with linear convex models" begin
-    systems = [c_sys5, c_sys14]
+    systems = [c_sys5] #c_sys14
     parameters_value = [true, false]
     networks = [DCPLLPowerModel,
                 LPACCPowerModel]
@@ -146,5 +149,77 @@ end
         psi_checksolve_test(UC, [MOI.OPTIMAL, MOI.LOCALLY_SOLVED], 340000, 100000)
         end
     end
+end
+################################################################
+duals = [:CopperPlateBalance]
+template = OperationsProblemTemplate(CopperPlatePowerModel, devices, branches, services);
+op_problem = OperationsProblem(TestOpProblem, template, c_sys5_re; optimizer = OSQP_optimizer, use_parameters = true)
+res = solve_op_problem!(op_problem; constraints_duals = duals)
 
+@testset "test constraint duals in the operations problem" begin
+    name = PSI.constraint_name("CopperPlateBalance")
+    for i in 1:ncol(res.time_stamp)
+        dual = JuMP.dual(op_problem.psi_container.constraints[name][i])
+        @test dual == res.constraints_duals[name][i, 1]
+    end
+end
+
+path = joinpath(pwd(), "test_writing")
+!isdir(path) && mkdir(path)
+
+function test_write_functions(file_path)
+
+    @testset "test _write_data functions" begin
+        PSI._write_data(res.variables, mkdir(joinpath(file_path, "one")))
+        readdir(joinpath(file_path, "one"))
+        for (k, v) in res.variables
+            @test isfile(joinpath(file_path, "one", "$k.feather"))
+        end
+
+        PSI._write_data(res.variables, res.time_stamp, mkdir(joinpath(file_path, "two")); file_type = CSV)
+        for (k, v) in res.variables
+            @test isfile(joinpath(file_path, "two/$k.csv"))
+        end
+
+        PSI._write_data(res.variables, res.time_stamp, mkdir(joinpath(file_path, "three")))
+        for (k, v) in res.variables
+            @test isfile(joinpath(file_path, "three", "$k.feather"))
+        end
+
+        var_name = PSI.variable_name(PSI.ACTIVE_POWER, PSY.ThermalStandard)
+        PSI._write_data(
+            res.variables[var_name],
+            mkdir(joinpath(file_path, "four")),
+            string(var_name)
+        )
+        @test isfile(joinpath(file_path, "four", "$(var_name).feather"))
+
+        #testing if directory is a file
+        PSI._write_data(
+            res.variables[var_name],
+            joinpath(file_path, "four", "$(var_name).feather"),
+            string(var_name)
+        )
+        @test isfile(joinpath(file_path, "four", "$(var_name).feather"))
+
+        PSI._write_optimizer_log(res.optimizer_log, mkdir(joinpath(file_path, "five")))
+        @test isfile(joinpath(file_path, "five", "optimizer_log.json"))
+
+        PSI.write_to_CSV(res, mkdir(joinpath(file_path, "six")))
+        @test !isempty(joinpath(file_path, "six", "results"))
+    end
+
+    @testset "test write result functions" begin
+        new_path = joinpath(file_path, "seven")
+        @test_throws IS.ConflictingInputsError PSI.write_results(res, new_path) # not yet a directory
+        PSI.write_results(res, mkdir(new_path))
+        @test !isempty(new_path)
+    end
+
+end
+
+try test_write_functions(path)
+finally
+    @info("removing test files")
+    rm(path, recursive=true)
 end
